app_vendas/requirements.txt
flask==2.2.3
werkzeug==2.2.3
psycopg2-binary==2.9.5
gunicorn==20.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0 